Use `_FILER_PATH` X11 atom to determine which Filer windows are open
Use `_FILER_PATH` atom to determine which Filer windows are open, instead of a separate registry which would occasionally get out of sync with reality. This should be more robust and follows our philosophy that data should be stored on the objects (e.g., windows, processes,...) themselves rather than in separate locations.

The downside of this change is that the current implementation will not work correctly with tabs. To make this work, we would need to store the paths of all tabs contained in a window in the X11 atom (rather than just the path of the frontmost tab), which would complicate things. Since we are not using tabs, this is not implemented.

This change modifies
https://github.com/helloSystem/Filer/commit/09aaaf44df876590b9e09c37bcd8eb59d18a30ac